Timers v0.1 AND PlayerController v2.0 AND CameraController v0.2 AND Graphics
Here comes a huge update today! Let's see what I've add.

Timers v0.1: The new class Timers let programmers set a timer specified by a string as the name of the timer. One can call Timers.SetTimer("exname", 0.5f) to start a timer named exname, and call Timers.isTimerFinished("exname") to check whether time's up. Also, call Timers.getTimerStart("exname") to get the gameTime as the timer is set. Notice that the timer doesn't change any variable in your class. The only thing it does is set the timer, and tell you if the time is out.

PlayerController v2.0: I remake the secondJump function. Now the second jump is not affected by the gravity, it just move to the mouse position in equal speed. After player arrives the destination, it stucks in the air for a short time, giving player time to kill the ball. THE NEXT UPDATE I will want to deal with the action that the player second jumps toward the ground, since the player should not be stuck in this case.

CameraController v0.2: The camera will not follow the player when it stucks in the air. It will lerp back to the player after the player finish stucking.

Graphics: I add the 特雷澤 sprite to the player. I will add animations to it in the future.