UNITY VERSION UPGRADE 2021.3.29f1 PlayerController v2.1 AND DirectionIndicator v0.1
I've upgraded the unity version of this project from 2021.3.20f1 to unity 2021.3.29f1, because the former has too many bugs.
Also, for the expected update in the previous commit, I successfully implemented one, but the animation part is put in the future. The next update will finish the saveReceive first.

PlayerController v2.1: Implements the message in the previous commit. Now when the player second jump under the top of the net, then it will not be stuck.

DirectionIndicator v0.1: A red line is drawn between the player and the mouse position. I want this line to help indicate the direction of jumping and upReceiving.