PlayerController v2.4 AND SegmentController vr1.0 AND BallActionController vr0.1 AND BallController vr0.1 AND Timers v0.3
First, I am sorry about my laziness and procrastination for committing the new changes. The last time I commit changes is at one month ago, and I have made a lot of changes since then, but I am too lazy to commit them. This causes the likelihood to lack some descriptions of this new update. I'll try my best to involve everything in the following description.

PlayerController v2.4: Add a new action called kill. When the player is flying high enough and is close to the ball, it can hold the mouse to charge (in the script I call this "killStuck"). When charging, the player and the ball pauses in the air. If it charges for enough time, the ball is killed. Otherwise, if the player release the mouse during the charging time, then the pause resumes.

SegmentController v0.1(original LineController): I altered the structure of the lines in this game. Now the SegmentController is the parent class. Children class should override the Update method to draw custom segments. For convenience, I created a method called Drawline(Vector3 a, Vector3 b) to draw a line from a to b so that childrens can call it in the Update. Also, to access the lineRenderer component, just use the variable lineRenderer, like what you did when using transform.

BallActionController v0.1(original BallController): I change the name of the class since it controls the action player do on the ball, instead of controlling the movement of the ball itself. I added some code to correspond to the kill action of PlayerController

BallController v0.1(original BallTest): Add the function PauseBall and ResumeBall.

Timers v0.3: Overload SetTimer(exTimer, duration, bool canOverride) to let user choose should the timer be reset when creating a timer that has existed already.